Working fullscreen with mouse position detection
It seems that mouse position detection varies a bit between browsers, but this works.  Also, I found a bug in Firefox in fullscreen mode when using personas.

It displays a small portion of the persona in full screen mode, and the size of the fullscreen element is affected, and mouse coordinates no longer match.  It also throws off the reported size of the window and makes it so that you can't assume your element is
screen size when in fullscreen mode.